task1/requirements.txt
Faker==37.4.0
psycopg2-binary==2.9.10
tzdata==2025.2
